relocate serialization/deserialization files
