+++ google-cloud-firestore/src/main/java/com/google/cloud/firestore/encoding/PojoBeanMapper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.firestore;
+package com.google.cloud.firestore.encoding;
 
 import com.google.cloud.Timestamp;
 import com.google.cloud.firestore.annotation.DocumentId;
